chore: rename existing mdxish test files
- take 1, theres a bit more to do
