+++ __tests__/lib/mdxish/mdxish-legacy-variables.test.ts
